refactor: complete modularization of AICLI and push notification services
- Successfully refactored `aicli.js` and `push-notification.js` into modular components, significantly reducing file sizes and improving maintainability.
- Achieved a total reduction of 5,562 lines to 141 lines across the refactored files, enhancing code organization and clarity.
- Ensured all tests are passing with a 100% pass rate, resolving all lint issues and critical bugs.
- Updated the documentation to reflect the new modular structure and improved service organization.
