+++ server/src/services/aicli-utils/index.js
@@ -1,0 +1,4 @@
+// Re-export all utilities from modular implementation
+export { InputValidator } from './input-validator.js';
+export { MessageProcessor } from './message-processor.js';
+export { AICLIConfig } from './aicli-config.js';
